requirements.txt
numpy
scipy
Pillow
scikit-image
